split tests more
