+++ packages/tests/tests/relay/utils.ts
@@ -1,5 +1,8 @@
-import { createDecoder, createEncoder } from "@waku/core";
+import { createDecoder, createEncoder, waitForRemotePeer } from "@waku/core";
+import { Protocols, RelayNode } from "@waku/interfaces";
+import debug from "debug";
 
+export const messageText = "Relay works!";
 export const TestContentTopic = "/test/1/waku-relay/utf8";
 export const TestEncoder = createEncoder({ contentTopic: TestContentTopic });
 export const TestDecoder = createDecoder(TestContentTopic);
@@ -13,3 +16,14 @@ export const CustomDecoder = createDecoder(
   CustomContentTopic,
   CustomPubSubTopic
 );
+
+export const log = debug("waku:test:relay");
+
+export async function waitForAllRemotePeers(
+  ...nodes: RelayNode[]
+): Promise<void> {
+  log("Wait for mutual pubsub subscription");
+  await Promise.all(
+    nodes.map((node) => waitForRemotePeer(node, [Protocols.Relay]))
+  );
+}
